Split processor, delegate and task files
